GF-7[feat]: 결제 요청 dto 생성
